+++ pages/products/index.vue
@@ -70,7 +70,7 @@
           <li>
             <a @click="$router.push('/slide')">Slide</a>
           </li>
-           <li>
+          <li>
             <a @click="$router.push('/comment')">Comment</a>
           </li>
           <li>
@@ -79,7 +79,7 @@
         </ul>
       </div>
       <div class="col-sm-8 col-lg-10 sidebar">
-        <b-button variant="success" @click="$router.push('/products/create')">Create new task</b-button>
+        <b-button style="margin-bottom:-130px" variant="success" @click="$router.push('/products/create')">Create new task</b-button>
         <div style="margin-left:1300px">
           <label>
             Sắp xếp theo
@@ -138,23 +138,19 @@
           :total-rows="pagination.total"
           :per-page="pagination.perPage"
           aria-controls="my-table"
-          @change="handleChange"
+          @change="handleChangePage"
         ></b-pagination>
       </div>
     </div>
-    <infinite-loading v-if="!search" @infinite="infiniteHandler"></infinite-loading>
   </div>
 </template>
 <script>
 const Cookie = process.client ? require("js-cookie") : undefined;
-import InfiniteLoading from "vue-infinite-loading";
 export default {
   mounted: function() {
     this.getTasks();
   },
-  components: {
-    InfiniteLoading
-  },
+
   data: function() {
     return {
       tasks: [],
@@ -172,25 +168,13 @@ export default {
       ],
       pagination: {
         currentPage: 1,
-        perPage: 15,
+        perPage: 10,
         totalPage: ""
-      }
+      },
+      cancel: false
     };
   },
   methods: {
-    infiniteHandler($state) {
-      setTimeout(() => {
-        this.pagination.currentPage++;
-        if (this.pagination.currentPage <= this.pagination.totalPage) {
-          this.getTasks();
-          setTimeout(() => {
-            $state.loaded();
-          }, 1000);
-        } else {
-          console.log("no more data");
-        }
-      }, 1000);
-    },
     getTasks() {
       let self = this;
       this.$axios
@@ -202,9 +186,8 @@ export default {
           this.order_by
         )
         .then(function(res) {
-          let temp = res.data.data;
-          self.tasks = self.tasks.concat(temp);
-          self.pagination.totalPage = res.data.pagination.totalPage;
+          self.tasks = res.data.data;
+          self.pagination = res.data.pagination;
         });
     },
     getProductsByOrder() {
@@ -235,10 +218,33 @@ export default {
       this.$router.push("/login");
     },
     delTasks: function(id) {
-      let self = this;
-      this.$axios.delete("/products/" + id).then(function(res) {
-        self.getTasks();
-      });
+      this.$swal
+        .fire({
+          title: "Bạn chắc chắn muốm xóa?",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonText: "Yes, delete it!",
+          cancelButtonText: "No, keep it"
+        })
+        .then(result => {
+          if (result.value) {
+            this.$swal.fire("Xóa!", "Bạn xóa sản phẩm thành công!.", "success");
+            let self = this;
+            this.$axios.delete("/products/" + id).then(function(res) {
+              self.getTasks();
+            });
+          } else if (result.dismiss === this.$swal.DismissReason.cancel) {
+            this.$swal.fire(
+              "Cancelled",
+              "Your imaginary file is safe :)",
+              "error"
+            );
+          }
+        });
+    },
+    handleChangePage(currentPage) {
+      this.pagination.currentPage = currentPage;
+      this.getTasks();
     }
   }
 };
